webapp/requirements.txt
colorama==0.3.3
cycler==0.10.0
Django==1.9.2
django-bootstrap3==6.2.2
django-model-utils==2.3.1
django-mptt==0.8.0
django-telegram-bot==0.5.0
djangorestframework==3.3.2
future==0.15.2
matplotlib==1.5.1
mysql-connector-python==2.1.3
numpy==1.10.4
Pillow==3.1.1
py==1.4.31
pyparsing==2.1.0
pyTelegramBotAPI==1.4.0
pytest==2.8.3
python-dateutil==2.5.0
python-slugify==1.1.4
python-telegram-bot==3.2.0
pytz==2015.7
requests==2.8.1
scipy==0.17.0
six==1.10.0
Unidecode==0.4.18
virtualenv==13.1.2
virtualenvwrapper-win==1.2.1
wheel==0.24.0
